+++ CHANGES.rst
@@ -10,6 +10,7 @@ vx.x.x (xxxx-xx-xx)
 - Add ``HashFS.folders()`` method that returns the folder paths that directly contain files (i.e. subpaths that only contain folders are ignored).
 - Add ``HashFS.detokenize()`` method that returns the file digest contained in a file path.
 - Add ``HashFS.repair()`` method that reindexes any files under root directory whose file path doesn't not match its tokenized file digest.
+- Rename ``Address.digest`` to ``Address.id``. (**breaking change**)
 
 
 v0.0.1 (2015-05-27)